refactor: 컴포넌트 페이지별 폴더 분리
