+++ src/pages/Cart.tsx
@@ -1,7 +1,7 @@
 import Header from "components/Header";
 import Page from "components/common/Page";
-import CartItemList from "components/CartItemList";
-import PurchaseOrder from "components/PurchaseOrder";
+import CartItemList from "components/cart/CartItemList";
+import PurchaseOrder from "components/cart/PurchaseOrder";
 import React from "react";
 import Skeleton from "components/common/Skeleton";
 import LoadingSpinner from "components/common/LoadingSpinner";
