+++ src/stories/components/CartItemList.stories.tsx
@@ -1,5 +1,5 @@
 import { StoryFn, Meta } from "@storybook/react";
-import CartItemList from "components/CartItemList";
+import CartItemList from "components/cart/CartItemList";
 
 export default {
   title: "CartItemList",
